+++ ava.config.js
@@ -1,5 +1,5 @@
 export default {
-  files: ["test/**/*.ts"],
+  files: ["test/**/*.test.ts"],
   extensions: ["ts"],
   require: ["./run-ts.js"],
   timeout: "2m",